Update Game.js
I forgot to commit my earlier changes. Implemented coloured pegs. Made it so pegs are deleted after being hit. Made it so 1 ball at a time can be fired from the top middle of the screen. Made it so bucket moves left and right and when ball hits it a life isn't removed. Life is deducted when a ball falls out of screen without hitting the bucket.